Created base framework for lighting system
Player class now sends out multiple Raycast2D nodes in all directions. This draws a Polygon2D around them (circle) when a ray intersects an obstacle, It change the point for the polygon. Essentially creating a field of sight that changes according to obstacles. Now just need to implement this with a lighting system, and prototype version should be 100% complete.